requirements.txt
djangorestframework
django-cors-headers
psycopg2-binary==2.8.3
django-environ